Move toy logic into its own folder
