+++ src/settings/components/ToyActuatorSettings.tsx
@@ -6,7 +6,7 @@ import {
   LinearMode,
   VibrateModeLabels,
   LinearModeLabels,
-} from '../../utils';
+} from '../../toy';
 import { PropsWithChildren, useState } from 'react';
 import { SettingsTile } from '../../common';
 import { SettingsDescription } from '../../common/SettingsDescription';
