+++ src/settings/components/ToyClientSettings.tsx
@@ -9,7 +9,8 @@ import {
   SettingsGrid,
   JoiStack,
 } from '../../common';
-import { defaultTransition, useToyClientValue, ToyClient } from '../../utils';
+import { defaultTransition } from '../../utils';
+import { useToyClientValue, ToyClient } from '../../toy';
 import {
   ButtplugBrowserWebsocketClientConnector,
   ButtplugClientDevice,
