+++ src/toy/index.ts
@@ -1,0 +1,2 @@
+export * from './toyactuator';
+export * from './toyclient';
